pyproject.toml
[project]
name = "recipe-rag-knowledge-graph"
version = "0.1.0"
description = "AI agent combining vector search and knowledge graph for recipe queries"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai>=0.3,<0.4",
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.34",
    "asyncpg>=0.30",
    "aiohttp>=3.0,<4.0",
    "graphiti-core>=0.12,<1.0",
    "neo4j>=5.28",
    "openai>=1.90",
    "python-dotenv>=1.1",
    "sse-starlette>=2.3",
    "pyyaml>=6.0",
    "prompt-toolkit>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4",
    "pytest-asyncio>=1.0",
    "black>=24.0",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "ingestion"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
